CVnatman/requirements.txt
tensorflow==2.13.0
pillow==10.0.1
numpy==1.24.3